Refactor: LoginPage
LoginPage.js를 Login.js로 변경하였습니다.
story를 추가하였습니다.
fontAwesomeIcon을 적용하여 LoginButton을 호출하도록 하였습니다.
