+++ frontend/src/index.js
@@ -5,7 +5,7 @@ import './index.css'
 import Channel from './page/channel/Channel'
 import { BrowserRouter, Route } from 'react-router-dom'
 import reportWebVitals from './reportWebVitals'
-import LoginPage from './page/LoginPage'
+import LoginPage from './page/login/Login'
 import WorkspaceSelectPage from './page/WorkspaceSelectPage'
 import Auth from './hooks/Auth'
 import GithubOAuth from './hooks/GithubOAuth'
