+++ frontend/src/page/login/Login.stories.js
@@ -1,0 +1,12 @@
+import React from 'react'
+import LoginPage from './Login'
+
+export default {
+  title: 'Example/Login',
+  component: LoginPage,
+}
+
+const Template = args => <LoginPage {...args} />
+
+export const Login = Template.bind({})
+Login.args = {}
